Address review round 3: CORS logging, KDF match cleanup, cert comment
- PgDeviceKeyRepo::create now wraps in explicit transaction (pool.begin)
  so FOR UPDATE lock is effective, not auto-commit
- Extract shared valid_signup_json() factory into tests/common/factories/signup.rs
- Remove duplicate valid_signup_body() from http_tests.rs and
  valid_signup_json() from identity_handler_tests.rs
- Fix clippy explicit_auto_deref in transaction usage
